feat+docs(v43): int16 (scale 10000) + Euclidean distance + v42 analysis
v42 RESULT (ADR-76): reproducibility confirmed
- FP: 734, FN: 412, score: +1075 (~= v38 +1082)
- Config K=5 thr=0.6 is stable and reproducible

CHAMPION ANALYSIS (ADR-77): MXLange/c-api-rinha2026
- int16 with scale 10000 (78x more precision than int8 0-127)
- Euclidean distance squared (not Manhattan)
- Tree partitioning by features + AVX2 SIMD + block pruning

v43 STRATEGY (ADR-78): int16 + Euclidean distance
- Vector14: [14]int8 -> [14]int16, sentinel -1
- Quantize: round(v * 10000) instead of round(v * 127)
- EuclideanDistanceSquared: sum(diff^2) instead of sum(|diff|)
- Loader: int16 LE serialization, magic IVF\x01->IVF\x02
- 84 MB vectors (was 42 MB), still fits in 165 MB

Hypothesis: int16 precision + L2 distance improves KNN ranking,
reducing FP+FN from ~1146 toward <800. Score target > +1500.

Changes:
- internal/vector/normalize.go: Vector14 int16, Quantize scale 10000,
  EuclideanDistanceSquared, ManhattanDistance kept for compat
- internal/vector/normalize_test.go: updated for int16 scale
- internal/index/index.go: ManhattanDistance -> EuclideanDistanceSquared
- internal/loader/loader.go: int16 LE serialization, magic IVF\x02
- internal/index/index_test.go: int16 centroid values
- DECISOES.md: ADR-76, ADR-77, ADR-78
- README.md: v42+v43, int16+Euclidean desc, benchmarks
- MELHORIAS.md: v42+v43 rows

+++ internal/index/index_test.go
@@ -19,7 +19,7 @@ func TestIVFSearch(t *testing.T) {
 	// Create centroids at increasing distances
 	for c := 0; c < nClusters; c++ {
 		for d := 0; d < 14; d++ {
-			centroids[c][d] = int8(c * 10)
+			centroids[c][d] = int16(c * 10)
 		}
 		offsets[c] = c * vectorsPerCluster
 	}
